Add LeaderOption
